requirements.txt
pandas
python-dotenv
streamlit
psycopg[binary]
openai
sqlalchemy